refactor: extract compareHttpInteraction to compare/oas/index to mirror asyncapi structure
